remove unnecessary type casting (diff_t should be long long)

+++ include/ppqsort/mainloop.h
@@ -23,7 +23,7 @@ namespace ppqsort::impl {
     template <typename RandomIt, typename diff_t = typename std::iterator_traits<RandomIt>::difference_type>
     inline void deterministic_shuffle(RandomIt begin, RandomIt end, const diff_t l_size, const diff_t r_size,
                                        RandomIt pivot_pos, const int insertion_threshold) {
-        if ((long long)l_size >= insertion_threshold) {
+        if (l_size >= insertion_threshold) {
             std::iter_swap(begin,             begin + l_size / 4);
             std::iter_swap(pivot_pos - 1, pivot_pos - l_size / 4);
 
@@ -35,7 +35,7 @@ namespace ppqsort::impl {
             }
         }
 
-        if ((long long)r_size >= insertion_threshold) {
+        if (r_size >= insertion_threshold) {
             std::iter_swap(pivot_pos + 1, pivot_pos + (1 + r_size / 4));
             std::iter_swap(end - 1,                   end - r_size / 4);
